+++ src/main.rs
@@ -42,6 +42,7 @@ fn sc_system(args: &ArgMatches) {
         Some(("make-orthogonal", s)) => sc_system_make_orthogonal(s),
         Some(("fix-permissions", s)) => sc_system_fix_permissions(s),
         Some(("forget", _)) => sc_system_forget(),
+        Some(("no-openmp", s)) => sc_system_no_openmp(s),
         _ => panic!("Usage: rim system [SUBCOMMAND], see help"),
     }
 }